+++ mobile/src/components/card/card.style.ts
