+++ mobile/src/components/card/card.type.ts
@@ -1,0 +1,6 @@
+import {Restaurant} from '../../types/data';
+
+export type TCardProps = {
+  restaurant?: Restaurant; // Optional for now, will be mandatory with data
+  //style props
+};
